Fix bug when rounding down bounded parameter values
The procedure that rounds down parameter values to integral values
ignored whether a parameter was bounded. As a result, it could round
down a parameter below its lower bound, which caused an error later on.
This commit does not round down parameters below their lower bound, if
they have one.