Improve handling of the case when there are no candidate probes
There may be some valid cases where there are no candidate probes. For
example, if genomes are clustered and --small-seq-skip is set, one
cluster may contain short genomes that are all skipped -- and thus and
the cluster has no genomes and therefore no candidate probes. This will
cause an error because kmer_probe_map would be empty. This commit allows
clusters/groupings of genomes to have no candidate probes, and for CATCH
to proceed in that case.