Change MinHash inner string hash function to md5
The universal hash function needs an inner hash function to hash strings
(i.e., k-mers). It had been using adler32 because Python's native
hash() is not determinsitic across sessions or processes, which leads
to problems when multiple processes use the same MinHash family.
However, adler32 was not a great choice (it is more for checksums than
hashing) because it has a high collision probability, which can lead to
unexpected behavior and difficult debugging. This commit changes that
inner hash function to be md5.

Since md5 is slower than adler32, this commit also adds an option
(use_fast_str_hash) to use Python's native hash() when deterministic
behavior is not needed for an instance of a family -- it can be ~10x
faster than md5.