Make default clustering method be simple search for connected components
This commit adds `utils.cluster.find_connected_components()` to identify
connected components in a graph in which sequences are adjacent iff
their estimated distance is within a threshold. As with previous
clustering, distances with this method continue to be estimated using a
MinHash family, e.g., as in the Mash paper (Ondov et al. 2016).

The commit makes this method the default for clustering because it
should be less memory-intensive than the existing hierarchical approach
for large numbers of input sequences; it does not need to store a
pairwise distance matrix. It adds an argument to design.py to specify
the clustering method, so the hierarchical clustering method---which
appears to be faster---can still be used.