Cover groupings separately by default
This removes the --cover-groupings-separately argument and instead
makes its behavior be the default behavior -- i.e., set_cover_filter
accepts groupings of target genomes and solves separate instances of
the set cover problem for each grouping, rather than one combined
instance. The bulk of changes are to set_cover_filter for this reason.

The other changes -- mostly to base_filter and probe_designer -- allow
probes to be passed to filters in a grouped manner. That is, the probes
are grouped to correspond to target genome groupings. Currently, only
set_cover_filter accepts grouped probes in that manner.