Parallelize filters over target groupings
This is only performed when the input (probes to filter) are grouped by
target genome groupings and the filter does not accept target genome
groupings. It is helpful for the filters in near_duplicate_filter, for
instance, which can be slow -- but it would not be applied by
set_cover_filter because the _filter() function there natively
parallelizes over target groupings already.