Fix bug in set_cover_filter with no candidate probes
There can be some cases where set_cover_filter is run with no input
candidate probes. For instance, if sequences are clustered and a
cluster/grouping has just 1 sequence that is both short (roughly the
length of a probe) and contains a string of Ns, then no candidate probes
will be constructed for it. This may happen sporadically because there
is randomness in the clustering algorithm.

In that case of a grouping with no input probes, set_cover_filter had
crashed. The call to
`probe.construct_kmer_probe_map_to_find_probe_covers()` on line 398 of
set_cover_filter.py returns an empty kmer_probe_map. Then, the call to
`probe.SharedKmerProbeMap.construct()` leads to a ValueError (line 708
of probe.py).

This commit fixes that case by checking for 0 candidate probes and
ensuring that the filter runs successfully, yielding 0 probes.