Add heuristic to decide which clustering method to use
The comment added by this commit to probe_designer details a problem
with the recently-introduced default method for clustering input
sequences ('simple'). The problem occurs on long genomes (e.g.,
bacterial) when also breaking the genomes into fragments and clustering
those.

This commit introduces a crude heuristic to select whether to use the
newer 'simple' method (which finds connected components) or the older,
most resource-intensive 'hierarchical' method (which uses agglomerative
hierarchical clustering). 'hierarchical' avoids the problem.

The argument --cluster-and-design-separately-method now has a new
choice, 'choose', to use the heuristic to decide the method, and
'choose' is the default setting for that argument.